refactor(user): relocate tanstack query code to presentation layer
Query keys and mutation options are presentation-only concerns; every caller already lives in presentation. queryKeys gains all()/session() (hierarchical, matching the shape used elsewhere in this refactor) and drops the unused userById key.
