--- src/module/user/infrastructure/tanstack/query/keys.ts
@@ -1,6 +1,0 @@
-const baseKey = 'user';
-
-export const queryKeys = {
-  sessionUser: [baseKey, 'session'],
-  userById: (id: string) => [baseKey, id] as const,
-} as const;
