+++ src/module/user/presentation/tanstack/query/keys.ts
@@ -1,0 +1,6 @@
+const baseKey = 'user';
+
+export const queryKeys = {
+  all: () => [baseKey] as const,
+  session: () => [...queryKeys.all(), 'session'] as const,
+} as const;
