+++ src/module/user/presentation/tanstack/query/options.ts
@@ -1,11 +1,11 @@
 import { queryOptions } from '@tanstack/react-query';
 
 import { userDataSource } from '@/user/dependency/data-source';
-import { queryKeys } from '@/user/infrastructure/tanstack/query/keys';
+import { queryKeys } from '@/user/presentation/tanstack/query/keys';
 
 export const getSessionUserQueryOptions = queryOptions({
   throwOnError: false,
-  queryKey: queryKeys.sessionUser,
+  queryKey: queryKeys.session(),
   queryFn: async () => {
     const userResult = await userDataSource.getSessionUser();
